backend/requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bakong-khqr==0.5.0
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.2.1
Eel==0.18.2
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
metatrader5==5.0.5488
moviepy==2.2.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pefile==2024.8.26
pillow==11.3.0
proglog==0.1.12
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyparsing==3.2.5
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
python-dotenv==1.2.1
python-vlc==3.0.21203
pytz==2025.2
pywin32-ctypes==0.2.3
qrcode==8.2
requests==2.32.5
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
starlette==0.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.40.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.5
zope.event==6.1
zope.interface==8.1.1
